Fix Sonar Issues

+++ legend-pure-m4/src/main/java/org/finos/legend/pure/m4/coreinstance/primitive/date/DateDiff.java
@@ -76,7 +76,7 @@ class DateDiff
         }
         else
         {
-            result = earlierLaterPair.getTwo().getOne().get(Calendar.DAY_OF_YEAR) - earlierLaterPair.getOne().getOne().get(Calendar.DAY_OF_YEAR);
+            result = (long)earlierLaterPair.getTwo().getOne().get(Calendar.DAY_OF_YEAR) - earlierLaterPair.getOne().getOne().get(Calendar.DAY_OF_YEAR);
         }
         return result;
     }
